refactor(chart): extract output/serialization methods into ChartOutputMixin
Move _repr_svg_, to_svg, to_markdown, to_html, _repr_html_, to_base64 and
save out of the Chart base class into a new ChartOutputMixin in
charted/charts/_chart_output.py. Behavior is unchanged: the mixin is added
as the first base so the methods still resolve on Chart via the MRO.
