+++ charted/charts/chart.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Protocol, cast
 
+from charted.charts._chart_output import ChartOutputMixin
 from charted.charts.axes import XAxis, YAxis
 from charted.constants import (
     AXIS_BORDER_COLOR,
@@ -20,10 +21,6 @@ from charted.themes.core import Theme
 from charted.utils.color_manager import ColorManager
 from charted.utils.data_model import DataModel
 from charted.utils.layout_engine import LayoutEngine
-from charted.utils.rendering import (
-    generate_html_wrapper,
-    generate_markdown_image,
-)
 from charted.utils.series_legend import SeriesLegend
 from charted.utils.theme_manager import ThemeManager
 from charted.utils.transform import translate
@@ -56,7 +53,7 @@ class _Annotation(Protocol):
     def render(self, chart: Chart) -> Element: ...
 
 
-class Chart(SeriesLegend, Svg):
+class Chart(ChartOutputMixin, SeriesLegend, Svg):
     """Base class for all SVG chart types.
 
     Provides common functionality for chart rendering including
@@ -98,102 +95,6 @@ class Chart(SeriesLegend, Svg):
     theme: Theme
     _title: MeasuredText | None
     _subtitle: MeasuredText | None
-
-    # =========================================================================
-    # Core Representation Methods
-    # =========================================================================
-
-    def _repr_svg_(self) -> str:
-        """Return SVG string for Jupyter notebook display."""
-        return self.svg
-
-    def to_svg(self) -> str:
-        """Get the SVG string representation of the chart."""
-        return self.svg
-
-    def to_markdown(self, alt_text: str | None = None, width: str | None = None) -> str:
-        """Generate markdown markup for the chart."""
-        title_text = self._title.text if self._title else None
-        return generate_markdown_image(self.svg, alt_text, title_text, width)
-
-    def to_html(
-        self, style: str = "display: inline-block;", tooltips: bool = False
-    ) -> str:
-        """Return standalone HTML with embedded SVG.
-
-        Args:
-            style: CSS style for the container div.
-            tooltips: If True, attach a native SVG ``<title>`` to each data
-                mark so browsers show a built-in hover tooltip (no
-                JavaScript). File output via ``to_svg()``/``save()`` is never
-                affected.
-
-        Returns:
-            HTML string with the SVG embedded in a div.
-        """
-        if not tooltips:
-            return generate_html_wrapper(self.svg, style)
-
-        # Regenerate the data-mark representation with <title> children, then
-        # restore the inert state so to_svg()/save() stay unchanged.
-        self._tooltips = True
-        try:
-            self._build_children()
-            svg = self.html
-        finally:
-            self._tooltips = False
-            self._build_children()
-        return generate_html_wrapper(svg, style)
-
-    def _repr_html_(self) -> str:
-        """Return HTML wrapper for the chart."""
-        return self.to_html()
-
-    def to_base64(self) -> str:
-        """Return the SVG as a data URI for inline embedding.
-
-        Returns:
-            Data URL string (e.g. 'data:image/svg+xml,<encoded-svg>').
-        """
-        from urllib.parse import quote
-
-        return f"data:image/svg+xml,{quote(self.svg)}"
-
-    def save(self, path: str, *, scale: int = 2) -> None:
-        """Save the chart to a file.
-
-        File format is detected from the extension:
-        - .svg: writes raw SVG markup
-        - .png: rasterizes via cairosvg (must be installed separately)
-
-        Args:
-            path: Destination file path.
-            scale: Resolution multiplier for PNG output (default 2x).
-
-        Raises:
-            ImportError: If saving as PNG and cairosvg is not installed.
-            ValueError: If the file extension is not supported.
-        """
-        import os
-
-        ext = os.path.splitext(path)[1].lower()
-
-        if ext == ".svg":
-            with open(path, "w") as f:
-                f.write(self.svg)
-        elif ext == ".png":
-            try:
-                import cairosvg  # type: ignore[import-untyped]
-            except ImportError:
-                raise ImportError(
-                    "PNG export requires cairosvg. "
-                    "Install it with: pip install cairosvg"
-                ) from None
-            cairosvg.svg2png(bytestring=self.svg.encode(), write_to=path, scale=scale)
-        else:
-            raise ValueError(
-                f"Unsupported file extension '{ext}'. Supported: .svg, .png"
-            )
 
     def style(self, **kwargs: object) -> "Chart":
         """Fluently apply theme overrides.
